refactor(web): split PlayerArea into focused subcomponents
Extract StockPile, HandSection and DiscardPiles (with their draggable
helpers) into apps/web/src/components/player-area/. PlayerArea.tsx
drops from 669 LOC to ~80 and re-exports the three pieces to keep
existing import sites working.
